Tour guide implementation: Implemented the suggested changes. (#539)
